more for folderItem

+++ src/main/java/org/example/technihongo/api/FlashCardController.java
@@ -1,9 +1,11 @@
 package org.example.technihongo.api;
 
+import org.example.technihongo.core.security.JwtUtil;
 import org.example.technihongo.dto.FlashcardRequestDTO;
 import org.example.technihongo.dto.FlashcardResponseDTO;
 import org.example.technihongo.response.ApiResponse;
 import org.example.technihongo.services.interfaces.FlashcardService;
+import org.example.technihongo.services.interfaces.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,18 +19,34 @@ public class FlashCardController {
     @Autowired
     private FlashcardService flashcardService;
 
-    @PostMapping("/{studentId}/{setId}/studentCreate")
+    @Autowired
+    private StudentService studentService;
+
+    @Autowired
+    private JwtUtil jwtUtil;
+
+
+    @PostMapping("/{setId}/studentCreate")
     public ResponseEntity<ApiResponse> createStudentFlashcards(
-            @PathVariable Integer studentId,
+            @RequestHeader("Authorization") String authorizationHeader,
             @PathVariable("setId") Integer flashcardSetId,
             @RequestBody List<FlashcardRequestDTO> requests) {
         try {
-            List<FlashcardResponseDTO> responses = flashcardService.createStudentFlashcards(studentId, flashcardSetId, requests);
-            return ResponseEntity.ok(ApiResponse.builder()
-                    .success(true)
-                    .message("Flashcards created successfully")
-                    .data(responses)
-                    .build());
+            if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+                String token = authorizationHeader.substring(7);
+                Integer userId = jwtUtil.extractUserId(token);
+                Integer studentId = studentService.getStudentIdByUserId(userId);
+
+                List<FlashcardResponseDTO> responseDTO = flashcardService.createStudentFlashcards(studentId, flashcardSetId, requests);
+                return ResponseEntity.ok(ApiResponse.builder()
+                        .success(true)
+                        .message("Flashcards created successfully")
+                        .data(responseDTO)
+                        .build());
+
+            } else {
+                throw new Exception("Authorization failed!");
+            }
         } catch (RuntimeException e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                     .body(ApiResponse.builder()
@@ -44,18 +62,25 @@ public class FlashCardController {
         }
     }
 
-    @PostMapping("/{userId}/{setId}/systemCreate")
+    @PostMapping("/{setId}/systemCreate")
     public ResponseEntity<ApiResponse> createSystemFlashcards(
-            @PathVariable Integer userId,
             @PathVariable("setId") Integer flashcardSetId,
-            @RequestBody List<FlashcardRequestDTO> requests) {
+            @RequestBody List<FlashcardRequestDTO> requests,
+            @RequestHeader("Authorization") String authorizationHeader) {
         try {
-            List<FlashcardResponseDTO> responses = flashcardService.createSystemFlashcards(userId, flashcardSetId, requests);
-            return ResponseEntity.ok(ApiResponse.builder()
-                    .success(true)
-                    .message("Flashcards created successfully")
-                    .data(responses)
-                    .build());
+            if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+                String token = authorizationHeader.substring(7);
+                Integer userId = jwtUtil.extractUserId(token);
+
+                List<FlashcardResponseDTO> responses = flashcardService.createSystemFlashcards(userId, flashcardSetId, requests);
+                return ResponseEntity.ok(ApiResponse.builder()
+                        .success(true)
+                        .message("Flashcards created successfully")
+                        .data(responses)
+                        .build());
+            } else {
+                throw new Exception("Authorization failed!");
+            }
         } catch (RuntimeException e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                     .body(ApiResponse.builder()
@@ -70,18 +95,25 @@ public class FlashCardController {
                             .build());
         }
     }
-    @PatchMapping("/{userId}/{flashcardId}/update")
+    @PatchMapping("/{flashcardId}/update")
     public ResponseEntity<ApiResponse> updateFlashcard(
-            @PathVariable Integer userId,
             @PathVariable Integer flashcardId,
-            @RequestBody FlashcardRequestDTO request) {
+            @RequestBody FlashcardRequestDTO request,
+            @RequestHeader("Authorization") String authorizationHeader) {
         try {
-            FlashcardResponseDTO response = flashcardService.updateFlashcard(userId, flashcardId, request);
-            return ResponseEntity.ok(ApiResponse.builder()
-                    .success(true)
-                    .message("Flashcard updated successfully")
-                    .data(response)
-                    .build());
+            if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+                String token = authorizationHeader.substring(7);
+                Integer userId = jwtUtil.extractUserId(token);
+
+                FlashcardResponseDTO response = flashcardService.updateFlashcard(userId, flashcardId, request);
+                return ResponseEntity.ok(ApiResponse.builder()
+                        .success(true)
+                        .message("Flashcard updated successfully")
+                        .data(response)
+                        .build());
+            } else {
+                throw new Exception("Authorization failed!");
+            }
         } catch (RuntimeException e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                     .body(ApiResponse.builder()
@@ -96,16 +128,23 @@ public class FlashCardController {
                             .build());
         }
     }
-    @DeleteMapping("/delete/{userId}/{flashcardId}")
+    @DeleteMapping("/delete/{flashcardId}")
     public ResponseEntity<ApiResponse> deleteFlashcard(
-            @PathVariable Integer userId,
-            @PathVariable Integer flashcardId) {
+            @PathVariable Integer flashcardId,
+            @RequestHeader("Authorization") String authorizationHeader) {
         try {
-            flashcardService.deleteFlashcard(userId, flashcardId);
-            return ResponseEntity.ok(ApiResponse.builder()
-                    .success(true)
-                    .message("Flashcard deleted successfully")
-                    .build());
+            if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+                String token = authorizationHeader.substring(7);
+                Integer userId = jwtUtil.extractUserId(token);
+
+                flashcardService.deleteFlashcard(userId, flashcardId);
+                return ResponseEntity.ok(ApiResponse.builder()
+                        .success(true)
+                        .message("Flashcard deleted successfully")
+                        .build());
+            } else {
+                throw new Exception("Authorization failed!");
+            }
         } catch (RuntimeException e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                     .body(ApiResponse.builder()
